+++ src/main/java/pl/asku/askumagazineservice/model/Reservation.java
@@ -5,6 +5,7 @@ import lombok.*;
 import javax.persistence.*;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Entity
@@ -34,7 +35,7 @@ public class Reservation {
     private LocalDate endDate;
 
     @NotNull
-    private Float areaInMeters;
+    private BigDecimal areaInMeters;
 
     @NotNull
     @ManyToOne(fetch = FetchType.LAZY)
